Move commit and rollback from PipelineJobPublicAPI to MigrationJobPublicAPI
